requirements.txt
pyyaml
sensehat
requests